refactor: organiza módulo DML em pacote próprio
- Cria pacote dml/ com estrutura modular
- Separa Assign em dml/assign.rs com testes próprios
- Separa Oop em dml/oop.rs com testes próprios
- Move DML enum para dml/mod.rs como coordenador do pacote
- Atualiza importações em genome.rs para usar a nova estrutura
- Mantém todos os testes funcionando (32 testes passando)
- Melhora organização e separação de responsabilidades no código
