+++ src/dsl/ast/mod.rs
@@ -3,6 +3,6 @@ pub mod fly;
 pub mod pog;
 pub mod genome;
 mod ddl;
-mod dml;
+pub mod dml;
 mod bug;
